requirements.txt
gym~=0.25
Pillow~=10.2
tqdm~=4.67
matplotlib~=3.10
numpy~=1.26
torch~=2.5
torchvision~=0.20
pandas~=2.2
moviepy~=1.0.3
tabulate~=0.9
safetensors~=0.5
pyarrow~=17.0
pytest~=8.3